Extract some code out of Camera2 file that is already massive
