+++ library/src/main/camera2/com/google/android/cameraview/Camera2.java
@@ -27,7 +27,6 @@ import android.hardware.camera2.CameraCharacteristics;
 import android.hardware.camera2.CameraDevice;
 import android.hardware.camera2.CameraManager;
 import android.hardware.camera2.CaptureRequest;
-import android.hardware.camera2.CaptureResult;
 import android.hardware.camera2.TotalCaptureResult;
 import android.hardware.camera2.params.StreamConfigurationMap;
 import android.media.Image;
@@ -78,8 +77,6 @@ class Camera2 extends CameraViewImpl {
      * Max preview height that is guaranteed by Camera2 API
      */
     private static final int MAX_PREVIEW_HEIGHT = 1080;
-
-    private static final int MAX_LOCK_FOCUS_ATTEMPTS = 3;
 
     private final CameraManager mCameraManager;
 
@@ -797,29 +794,11 @@ class Camera2 extends CameraViewImpl {
      * Updates the internal state of auto-focus to {@link #mAutoFocus}.
      */
     private void updateAutoFocus() {
-        mPreviewRequestBuilder.set(CaptureRequest.CONTROL_AF_MODE,
-                                   mAutoFocus ? pickAutoFocusMode() : CaptureRequest.CONTROL_AF_MODE_OFF);
-    }
-
-    private int pickAutoFocusMode() {
-        int[] supportedModes = mCameraCharacteristics.get(CameraCharacteristics.CONTROL_AF_AVAILABLE_MODES);
-        if (arrayContains(supportedModes, CaptureRequest.CONTROL_AF_MODE_CONTINUOUS_PICTURE)) {
-            // TODO Exclude S5 and Note 3 here.
-            return CaptureRequest.CONTROL_AF_MODE_CONTINUOUS_PICTURE;
-        } else if (arrayContains(supportedModes, CaptureRequest.CONTROL_AF_MODE_AUTO)) {
-            return CaptureRequest.CONTROL_AF_MODE_AUTO;
+        int afMode = CaptureRequest.CONTROL_AF_MODE_OFF;
+        if (mAutoFocus) {
+            afMode = FocusModeSelector.getBestAfMode(mCameraCharacteristics);
         }
-        return CaptureRequest.CONTROL_AF_MODE_OFF;
-    }
-
-    private boolean arrayContains(int[] array, int value) {
-        if (array == null) return false;
-        for (int i = 0; i < array.length; i++) {
-            if (array[i] == value) {
-                return true;
-            }
-        }
-        return false;
+        mPreviewRequestBuilder.set(CaptureRequest.CONTROL_AF_MODE, afMode);
     }
 
     /**
@@ -954,104 +933,6 @@ class Camera2 extends CameraViewImpl {
         } catch (CameraAccessException e) {
             Log.e(TAG, "Failed to restart camera preview.", e);
         }
-    }
-
-    /**
-     * A {@link CameraCaptureSession.CaptureCallback} for capturing a still picture.
-     */
-    private static abstract class PictureCaptureCallback
-            extends CameraCaptureSession.CaptureCallback {
-
-        public static final int STATE_PREVIEW = 0;
-        public static final int STATE_LOCKING = 1;
-        public static final int STATE_LOCKED = 2;
-        public static final int STATE_PRECAPTURE = 3;
-        public static final int STATE_WAITING = 4;
-        public static final int STATE_CAPTURING = 5;
-
-        private int mState;
-        private int mLockFocusAttemptsCount;
-
-
-        public void setState(int state) {
-            mState = state;
-        }
-
-        void clearLockFocusAttemptsCount() {
-            mLockFocusAttemptsCount = 0;
-        }
-
-        @Override
-        public void onCaptureProgressed(@NonNull CameraCaptureSession session,
-                                        @NonNull CaptureRequest request,
-                                        @NonNull CaptureResult partialResult) {
-            process(partialResult);
-        }
-
-        @Override
-        public void onCaptureCompleted(@NonNull CameraCaptureSession session,
-                                       @NonNull CaptureRequest request,
-                                       @NonNull TotalCaptureResult result) {
-            process(result);
-        }
-
-        private void process(@NonNull CaptureResult result) {
-            switch (mState) {
-                case STATE_LOCKING: {
-                    Integer af = result.get(CaptureResult.CONTROL_AF_STATE);
-                    if (af == null) {
-                        break;
-                    }
-                    Integer ae = result.get(CaptureResult.CONTROL_AE_STATE);
-                    boolean focused = af == CaptureResult.CONTROL_AF_STATE_FOCUSED_LOCKED;
-                    boolean notFocused = af == CaptureResult.CONTROL_AF_STATE_NOT_FOCUSED_LOCKED;
-                    if (focused || (notFocused && mLockFocusAttemptsCount >= MAX_LOCK_FOCUS_ATTEMPTS)) {
-                        if (ae == null || ae == CaptureResult.CONTROL_AE_STATE_CONVERGED) {
-                            setState(STATE_CAPTURING);
-                            onReady();
-                        } else {
-                            setState(STATE_LOCKED);
-                            onPrecaptureRequired();
-                        }
-                    } else if (notFocused) {
-                        onLockFocusRetryRequired();
-                        mLockFocusAttemptsCount++;
-                    }
-                    break;
-                }
-                case STATE_PRECAPTURE: {
-                    Integer ae = result.get(CaptureResult.CONTROL_AE_STATE);
-                    if (ae == null || ae == CaptureResult.CONTROL_AE_STATE_PRECAPTURE ||
-                            ae == CaptureRequest.CONTROL_AE_STATE_FLASH_REQUIRED ||
-                            ae == CaptureResult.CONTROL_AE_STATE_CONVERGED) {
-                        setState(STATE_WAITING);
-                    }
-                    break;
-                }
-                case STATE_WAITING: {
-                    Integer ae = result.get(CaptureResult.CONTROL_AE_STATE);
-                    if (ae == null || ae != CaptureResult.CONTROL_AE_STATE_PRECAPTURE) {
-                        setState(STATE_CAPTURING);
-                        onReady();
-                    }
-                    break;
-                }
-            }
-        }
-
-        /**
-         * Called when it is ready to take a still picture.
-         */
-        public abstract void onReady();
-
-        /**
-         * Called when it is necessary to run the precapture sequence.
-         */
-        public abstract void onPrecaptureRequired();
-        /**
-         * Called when locking focus has failed but MAX_LOCK_FOCUS_ATTEMPTS hasn't been reached yet
-         */
-        public abstract void onLockFocusRetryRequired();
     }
 
 }
